+++ tomcat/src/main/java/com/techcourse/service/UserService.java
@@ -3,9 +3,9 @@ package com.techcourse.service;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.UUID;
-import org.apache.coyote.http11.HttpRequestParameter;
-import org.apache.coyote.http11.Session;
-import org.apache.coyote.http11.SessionManager;
+import org.apache.coyote.http11.data.HttpRequestParameter;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
