+++ tomcat/src/main/java/org/apache/coyote/http11/data/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.data;
 
 public class HttpResponse {
     private HttpVersion httpVersion;
