+++ tomcat/src/main/java/org/apache/coyote/http11/data/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.data;
 
 import java.util.Arrays;
 import java.util.Optional;
