+++ tomcat/src/main/java/org/apache/coyote/http11/resource/ResourceReader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.resource;
 
 import java.io.BufferedReader;
 import java.io.File;
